fix(document): adapt to minio 9.0 stream() signature change
Minio 9.0 widened PutObjectArgs.stream() to accept (Long, Long); the
old (int, int) call no longer auto-boxes. Explicit casts keep the
single call site working.

+++ backend/services/document-service/src/main/java/cz/komercpoj/tmpmgmt/document/application/DocumentService.java
@@ -124,7 +124,7 @@ public class DocumentService {
     try {
       minio.putObject(
           PutObjectArgs.builder().bucket(minioProps.bucket()).object(key).stream(
-                  new ByteArrayInputStream(content), content.length, -1)
+                  new ByteArrayInputStream(content), (long) content.length, -1L)
               .contentType(format.mimeType())
               .build());
     } catch (Exception e) {
